# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Localiza Integração Cenário 4
  description: API demonstrativa para CRM, reservas e contratos, financeiro e faturamento,
    atendimento e assistência 24h, workflow e integração.
  version: 1.0.0
paths:
...
  /api/v1/crm/customers:
    post:
      tags:
      - CRM
      summary: Create Customer
      description: Cadastra cliente com identificador global e, opcionalmente, o identificador
        legado do CRM.
      operationId: create_customer_api_v1_crm_customers_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreate'
      responses:
        '201':
          description: Cliente criado no CRM com identificador global.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
              example:
                customerId: aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa
                name: Cliente Frota Localiza Ltda.
                email: gestor.frota@cliente-sintetico.example.com
                eligible: true
                consentService: true
                active: true
                legacyId: WEB-LOCALIZA-1001
        '401':
          description: Token ausente ou inválido.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:unauthorized
                title: Não autenticado
                status: 401
                detail: Token ausente ou inválido
                instance: /api/v1/crm/customers
                code: UNAUTHORIZED
                correlationId: 11111111-1111-4111-8111-111111111111
        '403':
          description: Papel sem permissão para a operação.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:forbidden
                title: Acesso negado
                status: 403
                detail: Papel sem permissão para esta operação
                instance: /api/v1/crm/customers
                code: FORBIDDEN
                correlationId: 11111111-1111-4111-8111-111111111111
        '409':
          description: E-mail ou identificador legado já cadastrado.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:conflict
                title: Conflito
                status: 409
                detail: E-mail ou identificador legado já cadastrado
                instance: /api/v1/crm/customers
                code: CONFLICT
                correlationId: 11111111-1111-4111-8111-111111111111
        '422':
          description: Entrada inválida.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:validation-error
                title: Entrada inválida
                status: 422
                detail:
                - *id001
                instance: /api/v1/crm/customers
                code: VALIDATION_ERROR
                correlationId: 11111111-1111-4111-8111-111111111111
                errors:
                - *id001
      parameters:
      - name: X-Correlation-ID
        in: header
        required: false
        description: UUID de correlação. A API gera um UUID quando o cabeçalho é omitido.
        schema:
          type: string
          format: uuid
        example: 11111111-1111-4111-8111-111111111111
    get:
      tags:
      - CRM
      summary: List Customers
      description: Lista clientes com filtros por e-mail, identificador legado e situação,
        com paginação.
      operationId: list_customers_api_v1_crm_customers_get
      security:
      - HTTPBearer: []
      parameters:
      - name: email
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
      - name: legacyId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Legacyid
      - name: active
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Active
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 200
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: X-Correlation-ID
        in: header
        required: false
        description: UUID de correlação. A API gera um UUID quando o cabeçalho é omitido.
        schema:
          type: string
          format: uuid
        example: 11111111-1111-4111-8111-111111111111
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerResponse'
                title: Response List Customers Api V1 Crm Customers Get
        '401':
          description: Token ausente ou inválido.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:unauthorized
                title: Não autenticado
                status: 401
                detail: Token ausente ou inválido
                instance: /api/v1/crm/customers
                code: UNAUTHORIZED
                correlationId: 11111111-1111-4111-8111-111111111111
        '403':
          description: Papel sem permissão para a operação.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:forbidden
                title: Acesso negado
                status: 403
                detail: Papel sem permissão para esta operação
                instance: /api/v1/crm/customers
                code: FORBIDDEN
                correlationId: 11111111-1111-4111-8111-111111111111
        '422':
          description: Entrada inválida.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:validation-error
                title: Entrada inválida
                status: 422
                detail:
                - *id001
                instance: /api/v1/crm/customers
                code: VALIDATION_ERROR
                correlationId: 11111111-1111-4111-8111-111111111111
                errors:
                - *id001
  /api/v1/crm/customers/{customer_id}:
    get:
      tags:
      - CRM
      summary: Get Customer
      description: Consulta um cliente pelo identificador global.
      operationId: get_customer_api_v1_crm_customers__customer_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: customer_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Customer Id
      - name: X-Correlation-ID
        in: header
        required: false
        description: UUID de correlação. A API gera um UUID quando o cabeçalho é omitido.
        schema:
          type: string
          format: uuid
        example: 11111111-1111-4111-8111-111111111111
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '401':
          description: Token ausente ou inválido.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:unauthorized
                title: Não autenticado
                status: 401
                detail: Token ausente ou inválido
                instance: /api/v1/crm/customers/{customer_id}
                code: UNAUTHORIZED
                correlationId: 11111111-1111-4111-8111-111111111111
        '403':
          description: Papel sem permissão para a operação.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:forbidden
                title: Acesso negado
                status: 403
                detail: Papel sem permissão para esta operação
                instance: /api/v1/crm/customers/{customer_id}
                code: FORBIDDEN
                correlationId: 11111111-1111-4111-8111-111111111111
        '404':
          description: Cliente não encontrado.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:not-found
                title: Recurso não encontrado
                status: 404
                detail: Cliente não encontrado
                instance: /api/v1/crm/customers/{customer_id}
                code: NOT_FOUND
                correlationId: 11111111-1111-4111-8111-111111111111
        '422':
          description: Entrada inválida.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:validation-error
                title: Entrada inválida
                status: 422
                detail:
                - *id001
                instance: /api/v1/crm/customers/{customer_id}
                code: VALIDATION_ERROR
                correlationId: 11111111-1111-4111-8111-111111111111
                errors:
                - *id001
    patch:
      tags:
      - CRM
      summary: Update Customer
      description: Altera nome, e-mail, elegibilidade ou consentimento; nome ou e-mail
        alterados publicam CustomerUpdated.v1.
      operationId: update_customer_api_v1_crm_customers__customer_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: customer_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Customer Id
      - name: X-Correlation-ID
        in: header
        required: false
        description: UUID de correlação. A API gera um UUID quando o cabeçalho é omitido.
        schema:
          type: string
          format: uuid
        example: 11111111-1111-4111-8111-111111111111
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '401':
          description: Token ausente ou inválido.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:unauthorized
                title: Não autenticado
                status: 401
                detail: Token ausente ou inválido
                instance: /api/v1/crm/customers/{customer_id}
                code: UNAUTHORIZED
                correlationId: 11111111-1111-4111-8111-111111111111
        '403':
          description: Papel sem permissão para a operação.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:forbidden
                title: Acesso negado
                status: 403
                detail: Papel sem permissão para esta operação
                instance: /api/v1/crm/customers/{customer_id}
                code: FORBIDDEN
                correlationId: 11111111-1111-4111-8111-111111111111
        '404':
          description: Cliente não encontrado.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:not-found
                title: Recurso não encontrado
                status: 404
                detail: Cliente não encontrado
                instance: /api/v1/crm/customers/{customer_id}
                code: NOT_FOUND
                correlationId: 11111111-1111-4111-8111-111111111111
        '409':
          description: E-mail já cadastrado para outro cliente.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:conflict
                title: Conflito
                status: 409
                detail: E-mail já cadastrado para outro cliente
                instance: /api/v1/crm/customers/{customer_id}
                code: CONFLICT
                correlationId: 11111111-1111-4111-8111-111111111111
        '422':
          description: Entrada inválida.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:validation-error
                title: Entrada inválida
                status: 422
                detail:
                - *id001
                instance: /api/v1/crm/customers/{customer_id}
                code: VALIDATION_ERROR
                correlationId: 11111111-1111-4111-8111-111111111111
                errors:
                - *id001
  /api/v1/crm/customers/{customer_id}/deactivate:
    post:
      tags:
      - CRM
      summary: Deactivate Customer
      description: Inativa o cliente sem apagar o histórico; cliente inativo não origina
        reserva nem contrato.
      operationId: deactivate_customer_api_v1_crm_customers__customer_id__deactivate_post
      security:
      - HTTPBearer: []
      parameters:
      - name: customer_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Customer Id
      - name: X-Correlation-ID
        in: header
        required: false
        description: UUID de correlação. A API gera um UUID quando o cabeçalho é omitido.
        schema:
          type: string
          format: uuid
        example: 11111111-1111-4111-8111-111111111111
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '401':
          description: Token ausente ou inválido.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:unauthorized
                title: Não autenticado
                status: 401
                detail: Token ausente ou inválido
                instance: /api/v1/crm/customers/{customer_id}/deactivate
                code: UNAUTHORIZED
                correlationId: 11111111-1111-4111-8111-111111111111
        '403':
          description: Papel sem permissão para a operação.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:forbidden
                title: Acesso negado
                status: 403
                detail: Papel sem permissão para esta operação
                instance: /api/v1/crm/customers/{customer_id}/deactivate
                code: FORBIDDEN
                correlationId: 11111111-1111-4111-8111-111111111111
        '404':
          description: Cliente não encontrado.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:not-found
                title: Recurso não encontrado
                status: 404
                detail: Cliente não encontrado
                instance: /api/v1/crm/customers/{customer_id}/deactivate
                code: NOT_FOUND
                correlationId: 11111111-1111-4111-8111-111111111111
        '422':
          description: Entrada inválida.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:validation-error
                title: Entrada inválida
                status: 422
                detail:
                - *id001
                instance: /api/v1/crm/customers/{customer_id}/deactivate
                code: VALIDATION_ERROR
                correlationId: 11111111-1111-4111-8111-111111111111
                errors:
                - *id001
...
components:
  schemas:
...
    CustomerCreate:
      properties:
        name:
          type: string
          maxLength: 150
          minLength: 2
          title: Name
        email:
          type: string
          format: email
          title: Email
        eligible:
          type: boolean
          title: Eligible
          default: true
        consentService:
          type: boolean
          title: Consentservice
          default: true
        legacyId:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Legacyid
      type: object
      required:
      - name
      - email
      title: CustomerCreate
      examples:
      - consentService: true
        eligible: true
        email: gestor.frota@cliente-sintetico.example.com
        legacyId: WEB-LOCALIZA-1001
        name: Cliente Frota Localiza Ltda.
    CustomerResponse:
      properties:
        customerId:
          type: string
          format: uuid
          title: Customerid
        name:
          type: string
          title: Name
        email:
          type: string
          format: email
          title: Email
        eligible:
          type: boolean
          title: Eligible
        consentService:
          type: boolean
          title: Consentservice
        active:
          type: boolean
          title: Active
          default: true
        legacyId:
          anyOf:
          - type: string
          - type: 'null'
          title: Legacyid
      type: object
      required:
      - customerId
      - name
      - email
      - eligible
      - consentService
      title: CustomerResponse
      examples:
      - active: true
        consentService: true
        customerId: aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa
        eligible: true
        email: gestor.frota@cliente-sintetico.example.com
        legacyId: WEB-LOCALIZA-1001
        name: Cliente Frota Localiza Ltda.
    CustomerUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 150
            minLength: 2
          - type: 'null'
          title: Name
        email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
        eligible:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Eligible
        consentService:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Consentservice
      type: object
      title: CustomerUpdate
      examples:
      - consentService: true
        name: Cliente Frota Sintética Atualizada
...
    ProblemDetails:
      description: Corpo de erro publicado no OpenAPI.
      examples:
      - code: NOT_FOUND
        correlationId: 11111111-1111-4111-8111-111111111111
        detail: Contrato não encontrado
        instance: /api/v1/contracts/bbbbbbbb-bbbb-4bbb-8bbb-bbbbbbbbbbbb
        status: 404
        title: Recurso não encontrado
        type: urn:archcorp:problem:not-found
      properties:
        type:
          description: URI que identifica o tipo do problema.
          title: Type
          type: string
        title:
          title: Title
          type: string
        status:
          title: Status
          type: integer
        detail:
          anyOf:
          - type: string
          - items:
              additionalProperties: true
              type: object
            type: array
          description: Mensagem legível. Em 422 mantém a lista de erros da versão
            1.
          title: Detail
        instance:
          description: Caminho da requisição que originou o erro.
          title: Instance
          type: string
        code:
          description: Código estável para tratamento pelo cliente.
          title: Code
          type: string
        correlationId:
          title: Correlationid
          type: string
        errors:
          anyOf:
          - items:
              additionalProperties: true
              type: object
            type: array
          - type: 'null'
          default: null
          title: Errors
      required:
      - type
      - title
      - status
      - detail
      - instance
      - code
      - correlationId
      title: ProblemDetails
      type: object
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer